separate command and click modules
